refactor worker classes
