+++ alfalfa_worker/worker_fmu/worker.py
@@ -26,10 +26,6 @@
 # -*- coding: utf-8 -*-
 from __future__ import print_function
 
-import os
-import subprocess
-import sys
-
 from alfalfa_worker.worker_job_base import WorkerJobBase
 
 
@@ -40,7 +36,7 @@ class WorkerFmu(WorkerJobBase):
     def __init__(self):
         super().__init__()
 
-    def check_step_sim_config(self, message_body):
+    def _check_step_sim_config(self, message_body):
         """
         Check that the configuration parameters sent in the message for step_sim are valid.
         Message body must have exactly one of:
@@ -61,75 +57,11 @@ class WorkerFmu(WorkerJobBase):
         start = "0"
         end = str(60 * 60 * 24 * 365)
 
-        # TODO change server side message: startDatetime to start_datetime
-        start_datetime = message_body.get('startDatetime', start)
-        if start_datetime == 'undefined':
-            start_datetime = start
-        # TODO change server side message: endDatetime to end_datetime
-        end_datetime = message_body.get('endDatetime', end)
-        if end_datetime == 'undefined':
-            end_datetime = end
-
-        realtime = message_body.get('realtime', False) == 'true'
-        timescale = message_body.get('timescale', False)  # type checking below in if statement
-        timescale = False if timescale == 'undefined' else timescale
-        # TODO change server side message: externalClock to external_clock
-        external_clock = message_body.get('externalClock', False) == 'true'
-
-        # TODO remove after tests written
-        self.logger.info(
-            "start_datetime type: {}\tstart_datetime: {}".format(type(start_datetime), start_datetime))
-        self.logger.info(
-            "end_datetime type: {}\tend_datetime: {}".format(type(end_datetime), end_datetime))
-        self.logger.info("realtime type: {}\trealtime: {}".format(type(realtime), realtime))
-        self.logger.info("timescale type: {}\ttimescale: {}".format(type(timescale), timescale))
-        self.logger.info(
-            "external_clock type: {}\texternal_clock: {}".format(type(external_clock), external_clock))
-        # Only want one of: realtime, timescale, or external_clock.  Else, reject configuration.
-        # if (realtime and timescale) or (realtime and external_clock) or (timescale and external_clock):
-        #    self.logger.info(
-        #        "Only one of 'external_clock', 'timescale', or 'realtime' should be specified in message")
-        #    sys.exit(1)
-
-        # Check for at least one of the required parameters
-        if not realtime and not timescale and not external_clock:
-            # TODO: If exiting, what logging type to use?
-            self.logger.info(
-                "At least one of 'external_clock', 'timescale', or 'realtime' must be specified")
-            sys.exit(1)
-
-        if external_clock:
-            step_sim_type = "external_clock"
-            step_sim_value = "true"
-        elif realtime:
-            step_sim_type = "realtime"
-            step_sim_value = 1
-        elif timescale:
-            if str(timescale).isdigit():
-                step_sim_type = "timescale"
-                step_sim_value = int(timescale)
-            else:
-                self.logger.info(f"timescale: {timescale} must be an integer value")
-                sys.exit(1)
+        (step_sim_type, step_sim_value, start_datetime, end_datetime) = self._check_step_sim_time_config(message_body, start, end)
 
         return (step_sim_type, step_sim_value, start_datetime, end_datetime)
 
-    def add_site_type(self, p, file_name, upload_id):
-        """
-        Simple wrapper for the add_site subprocess call given the path for python file to call
-
-        :param p: path of add_site/add_site.py file to call
-        :param file_name: name of file to add with extension, excludes full path
-        :param upload_id:
-        :return:
-        """
-        if not os.path.isfile(p):
-            self.logger.info(f"No file: {p}, Working directory: {os.getcwd()}")
-        else:
-            return_code = subprocess.call(['python3', p, file_name, upload_id])
-            self.check_subprocess_call(return_code, file_name, 'add_site')
-
-    def step_sim_type(self, site_id, step_sim_type, step_sim_value, start_datetime, end_datetime, model_type):
+    def _call_step_sim_process(self, site_id, step_sim_type, step_sim_value, start_datetime, end_datetime):
         """Simple wrapper for the step_sim subprocess call given required params
 
         :param p:
@@ -138,130 +70,48 @@ class WorkerFmu(WorkerJobBase):
         :param step_sim_value:
         :param start_datetime:
         :param end_datetime:
-        :param model_type: string, type of model. Choices are 'fmu'
         :return:
         """
-        self.logger.info("Calling model '{}' step_sim_type '{}'".format(model_type, step_sim_type))
+        self.logger.info("Calling model '{}' step_sim_type 'fmu'".format(step_sim_type))
         arg_step_sim_value = None
-        if step_sim_type == 'external_clock':
-            p = 'alfalfa_worker/worker_fmu/step_fmu.py'
-            # fmu needs to run with Python 2
-            python = 'python'
-            arg_start_datetime = start_datetime
-            arg_end_datetime = end_datetime
-        else:
-            p = 'alfalfa_worker/worker_fmu/step_fmu.py'
-            # fmu needs to run with Python 2
-            python = 'python'
+
+        p = 'alfalfa_worker/worker_fmu/step_fmu.py'
+        # fmu needs to run with Python 2
+        python = 'python'
+        arg_start_datetime = start_datetime
+        arg_end_datetime = end_datetime
+        if step_sim_type != 'external_clock':
             arg_step_sim_value = step_sim_value
-            arg_start_datetime = start_datetime
-            arg_end_datetime = end_datetime
 
-        if os.path.isfile(p):
-            call = [
-                python, p, site_id, step_sim_type, arg_start_datetime, arg_end_datetime
-            ]
-            if arg_step_sim_value:
-                call.append('--step_sim_value={}'.format(step_sim_value))
-
-            self.logger.info("Calling step_sim_type subprocess: {}".format(call))
-            return_code = subprocess.call(call)
-
-            self.check_subprocess_call(return_code, site_id, 'step_sim')
-        else:
-            self.logger.info("No file: {}".format(p))
-            sys.exit(1)
+        args = [site_id, step_sim_type, arg_start_datetime, arg_end_datetime]
+        if arg_step_sim_value:
+            args.append('--step_sim_value={}'.format(step_sim_value))
+        self.logger.info("Calling step_sim_type subprocess: {}".format([python, p] + args))
+        return_code = self._call_subprocess(python, p, args)
+        self._check_subprocess_call(return_code, site_id, 'step_sim')
 
         return return_code
 
-    def run_sim_type(self, p, file_name, upload_id):
+    def _call_run_sim_process(self, file_name, upload_id):
         """
         Simple wrapper for the run_sim subprocess call given the path for python file to call
 
-        :param p: path of file to call
         :param file_name: name of file to run with extension
         :param upload_id:
         :return:
         """
-        if not os.path.isfile(p):
-            self.logger.info("No file: {}".format(p))
-        else:
-            return_code = subprocess.call(['python3', p, file_name, upload_id])
-            self.check_subprocess_call(return_code, file_name, 'run_sim')
+        p = 'run_sim/sim_fmu/sim_fmu.py'
+        return_code = self._call_subprocess('python3', p, [file_name, upload_id])
+        self._check_subprocess_call(return_code, file_name, 'run_sim')
 
-    def add_site(self, message_body):
+    def _call_add_site_process(self, file_name, upload_id):
         """
-        Caller to add a site depending on the type of file to upload.
-        Valid file extensions include: '.fmu'
-        Site is added by spinning off a subprocess
-        TODO: Document expected directory structure for OS workflows
+        Simple wrapper for the add_site subprocess call given the path for python file to call
 
-        :param message_body: Body of a single message from a boto3 Queue resource
-        :type message_body: dict
+        :param file_name: name of file to add with extension, excludes full path
+        :param upload_id:
         :return:
         """
-        # TODO: move the body check to before the calling of `add_site`
-        body_check = self.check_message_body(message_body, 'add_site')
-        if not body_check:
-            self.logger.info(
-                'message_body for add_site does not have correct keys.  Site will not be added.')
-            # TODO insert sys.exit(1)?
-        else:
-            file_name = message_body.get('model_name')
-            upload_id = message_body.get('upload_id')
-            self.logger.info('Add site for file_name: %s, and upload_id: %s' % (file_name, upload_id))
-
-            p = 'alfalfa_worker/worker_fmu/add_site.py'
-            self.add_site_type(p, file_name, upload_id)
-
-    def step_sim(self, message_body):
-        """
-        Caller to step through a simulation.
-        Valid simulation types include: '.fmu'
-        The simulation will not be started if an invalid simulation config is provided.
-        Run occurs by spinning of a subprocess.
-
-        :param message_body: Body of a single message from a boto3 Queue resource
-        :type message_body: dict
-        :return:
-        """
-        body_check = self.check_message_body(message_body, 'step_sim')
-        if not body_check:
-            self.logger.info(
-                'message_body for step_sim does not have correct keys.  Simulation will not be stepped through.')
-            # TODO insert sys.exit(1)?
-        else:
-            site_id = message_body.get('id')
-            site_rec = self.mongo_db_recs.find_one({"_id": site_id})
-            self.logger.info('Site record: {}'.format(site_rec))
-
-            # TODO: Check if simType is defined, error if not.
-            sim_type = site_rec.get("rec", {}).get("simType", "fmu").replace("s:", "")
-            step_sim_type, step_sim_value, start_datetime, end_datetime = self.check_step_sim_config(message_body)
-            self.logger.info('Start step_sim for site_id: %s, and sim_type: %s' % (site_id, sim_type))
-
-            # TODO: do we need to have two versions of python?
-            self.step_sim_type(site_id, step_sim_type, step_sim_value, start_datetime, end_datetime, sim_type)
-
-    def run_sim(self, message_body):
-        """
-        Run a simulation.  Running a simulation provides vanilla run of either an OSM or an FMU model.  Doesn't
-        allow for external interfacing into model during runtime - simply runs the model.  Run occurs
-        by spinning off a subprocess.
-
-        :param message_body: Body of a single message from a boto3 Queue resource
-        :type message_body: dict
-        :return:
-        """
-        body_check = self.check_message_body(message_body, 'run_sim')
-        if not body_check:
-            self.logger.info(
-                'message_body for step_sim does not have correct keys.  Simulation will not be run.')
-        else:
-            upload_filename = message_body.get('upload_filename')
-            upload_id = message_body.get('upload_id')
-            self.logger.info(
-                'Run sim for upload_filename: %s, and upload_id: %s' % (upload_filename, upload_id))
-
-            p = 'run_sim/sim_fmu/sim_fmu.py'
-            self.run_sim_type(p, upload_filename, upload_id)
+        p = 'alfalfa_worker/worker_fmu/add_site.py'
+        return_code = self._call_subprocess('python3', p, [file_name, upload_id])
+        self._check_subprocess_call(return_code, file_name, 'add_site')
